Include generated files in build, abort build if tests fail on Win32, use double for sub-millisecond accuracy on timing tests, provide default standard functions so "release" client tests pass.

+++ projects/OG-Language/ConnectorTest/OverlappedTest.cpp
@@ -20,8 +20,8 @@ LOGGING (com.opengamma.language.connector.OverlappedTest);
 #define RETRIES				10
 
 static CConnector *g_poConnector;
-static int g_tOverlappedQuery = 0;
-static int g_tInlineQuery = 0;
+static double g_tOverlappedQuery = 0;
+static double g_tInlineQuery = 0;
 
 static void StartConnector () {
 	g_poConnector = CConnector::Start (TEST_LANGUAGE);
